refactor(phase-9): split models.py into focused modules
Convert ``src/services/intelligence/models.py`` (single 500-line file
covering five concerns) into a package of focused modules:

- ``models/graph.py``       — NodeType, EdgeType, GraphNode, GraphEdge,
                               ENTITY_NAME_PATTERN
- ``models/knowledge.py``   — EntityType, ExtractedEntity,
                               ExtractedRelation
- ``models/frontier.py``    — TrendStatus, GapType
- ``models/monitoring.py``  — PaperSource, SubscriptionLimitError
- ``models/exceptions.py``  — GraphStoreError + subclasses,
                               OptimisticLockError

The package ``__init__`` re-exports ONLY the shared kernel (graph
models + exceptions). Milestone-specific symbols (entities, trends,
sources, etc.) must be imported from their submodules so each milestone
opts into exactly the surface it needs. The intelligence package
``__init__`` already re-exports just the kernel and is unchanged.

Decision (item 15, DTO vs. Optional): keep ``version``/``created_at``/
``updated_at`` as non-Optional with auto-generating defaults. The
storage layer is the single owner of these fields and overwrites them
on every read; introducing a separate persistence DTO would churn every
downstream consumer for negligible safety gain (Pydantic strict still
enforces type on construction). Documented inline in models/graph.py.

Tests: 4232 passing. Coverage: 99.18%.
